backend/requirements.txt
# Python dependencies for backend with pinned versions for reproducibility
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
stripe==7.9.0
python-dotenv==1.0.0
Flask-WTF==1.2.1
Werkzeug==2.3.7
SQLAlchemy==2.0.23
Jinja2==3.1.2
# Security packages
pyjwt==2.8.0
itsdangerous==2.1.2
# Testing and quality
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
black==23.10.0
